+++ src/st/st0/36358.c
@@ -265,6 +265,7 @@ void func_801B6C5C(Entity* entity) {
 
 INCLUDE_ASM("st/st0/nonmatchings/36358", EntityEquipItemDrop);
 
+// FIXME: split this file to remove the need to pad this string
 const char D_801A7CC4[] = "\202\360\223\374\216\350\202\265\202\275\000\000";
 
 INCLUDE_ASM("st/st0/nonmatchings/36358", func_801B7308);
